chore: extract useUserSession internals, dedupe isRecord, merge tests
- Split useUserSession.ts (518→260L) into redirect-helpers, session-fetch, wrap-auth-method
- Extract shared isRecord() to internal/utils.ts (was duplicated 3x)
- Merge secondary storage unit tests into single file
- Remove stale agents field and skills from package.json
